pyproject.toml
[project]
name = "antislopfactory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "datasets>=4.5.0",
    "fastapi[standard]>=0.129.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
]
